backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
pdf2image==1.16.3
Werkzeug==2.3.7
Pillow==10.0.0